# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "crossbill API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/books/{book_id}/digest": {
      "get": {
        "tags": [
          "digest"
        ],
        "summary": "Get Book Digest",
        "description": "Get the digest of every chapter in a book.",
        "operationId": "get_book_digest",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse_ChapterDigestResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/books/{book_id}/digest": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Enqueue Book Digest",
        "description": "Enqueue digest generation for all chapters of a book.",
        "operationId": "enqueue_book_digest",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Active Book Digest Batch",
        "description": "Get the active digest batch for a book, if any.",
        "operationId": "get_active_book_digest_batch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/JobBatchResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Active Book Digest Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/batches/{batch_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Batch",
        "description": "Get job batch status.",
        "operationId": "get_job_batch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job Batch",
        "description": "Cancel a job batch and abort all pending/active jobs.",
        "operationId": "cancel_job_batch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/semantic/backfill": {
      "post": {
        "tags": [
          "semantic"
        ],
        "summary": "Backfill Embeddings",
        "description": "Enqueue an embedding batch for the user's content, optionally scoped to a book.\n\nAnswers 202 with the batch to poll through ``GET /jobs/batches/{id}``, or 200\nwith ``total_jobs`` 0 when everything is already indexed. The second is a\nnormal outcome of pressing backfill twice, which is why it is not a 4xx: a\ncaller cannot tell a rejected request from a satisfied one.\n\n409 when a backfill is already running -- ``GET /semantic/backfill/active``\nnames it, and cancelling it clears the way.",
        "operationId": "backfill_embeddings",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "book_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillResponse"
                }
              }
            }
          },
          "200": {
            "description": "Everything was already indexed; no batch was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackfillResponse"
                }
              }
            }
          },
          "409": {
            "description": "A backfill is already running; cancel it or wait"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/semantic/backfill/active": {
      "get": {
        "tags": [
          "semantic"
        ],
        "summary": "Get Active Backfill",
        "description": "Get the user's running backfill, if any -- what a 409 from POST refers to.",
        "operationId": "get_active_backfill",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/JobBatchResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Active Backfill"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/semantic/search": {
      "get": {
        "tags": [
          "semantic"
        ],
        "summary": "Search Content",
        "description": "Rank the user's embedded content by semantic similarity to a free-text query.",
        "operationId": "search_content",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1000,
              "title": "Q"
            }
          },
          {
            "name": "book_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Book Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SemanticSearchResult"
                  },
                  "title": "Response Search Content"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/semantic/related": {
      "get": {
        "tags": [
          "semantic"
        ],
        "summary": "Related Content",
        "description": "Rank the user's embedded content by similarity to one already-indexed unit.",
        "operationId": "related_content",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "content_type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ContentType"
            }
          },
          {
            "name": "content_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Content Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SemanticSearchResult"
                  },
                  "title": "Response Related Content"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackfillResponse": {
        "properties": {
          "total_jobs": {
            "type": "integer",
            "title": "Total Jobs"
          },
          "batch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobBatchResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "total_jobs",
          "batch"
        ],
        "title": "BackfillResponse",
        "description": "What a backfill request did.\n\n``total_jobs`` is the field to key on, because it is the only one present in\nboth outcomes: it is 0 exactly when nothing needed embedding, and then\n``batch`` is null because no batch was created and there is nothing to poll."
      },
...
      "ChapterDigestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "chapter_id": {
            "type": "integer",
            "title": "Chapter Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "keypoints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keypoints"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/DigestQuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "chapter_id",
          "summary",
          "keypoints",
          "questions",
          "generated_at"
        ],
        "title": "ChapterDigestResponse",
        "description": "Response schema for a chapter digest."
      },
...
      "CollectionResponse_ChapterDigestResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChapterDigestResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CollectionResponse[ChapterDigestResponse]"
      },
...
      "ContentType": {
        "type": "string",
        "enum": [
          "note",
          "highlight",
          "digest"
        ],
        "title": "ContentType",
        "description": "A kind of natural-language unit that can be embedded."
      },
...
      "DigestQuestionResponse": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          }
        },
        "type": "object",
        "required": [
          "question",
          "answer",
          "user_answer"
        ],
        "title": "DigestQuestionResponse",
        "description": "Response schema for a digest question/answer pair."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobBatchResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "batch_type": {
            "type": "string",
            "title": "Batch Type"
          },
          "reference_id": {
            "type": "string",
            "title": "Reference Id"
          },
          "total_jobs": {
            "type": "integer",
            "title": "Total Jobs"
          },
          "completed_jobs": {
            "type": "integer",
            "title": "Completed Jobs"
          },
          "failed_jobs": {
            "type": "integer",
            "title": "Failed Jobs"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_type",
          "reference_id",
          "total_jobs",
          "completed_jobs",
          "failed_jobs",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobBatchResponse"
      },
...
      "SemanticSearchResult": {
        "properties": {
          "content_type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "content_id": {
            "type": "integer",
            "title": "Content Id"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "content_type",
          "content_id",
          "book_id",
          "score",
          "text"
        ],
        "title": "SemanticSearchResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}